fix project settings contract boundary
