+++ packages/project-build/src/contracts.ts
@@ -4,3 +4,4 @@ export * from "./contracts/input-schema";
 export * from "./contracts/namespaces";
 export * from "./contracts/permissions";
 export * from "./contracts/patch";
+export * from "./contracts/project-settings";
